+++ packages/mrm-task-ci/CHANGELOG.md
@@ -3,6 +3,33 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.2.3](https://github.com/sapegin/mrm/compare/mrm-task-ci@0.2.2...mrm-task-ci@0.2.3) (2021-02-08)
+
+**Note:** Version bump only for package mrm-task-ci
+
+
+
+
+
+## [0.2.2](https://github.com/sapegin/mrm/compare/mrm-task-ci@0.2.1...mrm-task-ci@0.2.2) (2021-02-03)
+
+**Note:** Version bump only for package mrm-task-ci
+
+
+
+
+
+## [0.2.1](https://github.com/sapegin/mrm/compare/mrm-task-ci@0.2.0...mrm-task-ci@0.2.1) (2020-11-25)
+
+
+### Bug Fixes
+
+* **ci:** Detect default Git branch name ([f03647e](https://github.com/sapegin/mrm/commit/f03647ece853d0acd2afe142da716568249a9fe9))
+
+
+
+
+
 # [0.2.0](https://github.com/sapegin/mrm/compare/mrm-task-ci@0.1.1...mrm-task-ci@0.2.0) (2020-11-05)