+++ packages/mrm-task-codecov/CHANGELOG.md
@@ -3,6 +3,38 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [3.0.2](https://github.com/sapegin/mrm/compare/mrm-task-codecov@3.0.1...mrm-task-codecov@3.0.2) (2021-02-08)
+
+**Note:** Version bump only for package mrm-task-codecov
+
+
+
+
+
+## [3.0.1](https://github.com/sapegin/mrm/compare/mrm-task-codecov@3.0.0...mrm-task-codecov@3.0.1) (2021-02-03)
+
+**Note:** Version bump only for package mrm-task-codecov
+
+
+
+
+
+# [3.0.0](https://github.com/sapegin/mrm/compare/mrm-task-codecov@2.1.3...mrm-task-codecov@3.0.0) (2020-11-25)
+
+
+### Features
+
+* **codecov:** Create a GitHub Actions workflow instead of Travis CI ([65a442b](https://github.com/sapegin/mrm/commit/65a442b2ee654a1d63a08f4173a4185359714464))
+
+
+### BREAKING CHANGES
+
+* **codecov:** Create a GitHub Actions workflow instead of Travis CI
+
+
+
+
+
 ## [2.1.3](https://github.com/sapegin/mrm/compare/mrm-task-codecov@2.1.2...mrm-task-codecov@2.1.3) (2020-11-04)
 
 **Note:** Version bump only for package mrm-task-codecov